requirements.txt
aiohttp==3.8.1
click==8.0.3
python-dotenv==0.19.2
